JVMCBC-315: Use double for ack threshold in Builder
Reviewed-on: http://review.couchbase.org/61925
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>

+++ src/main/java/com/couchbase/client/core/env/DefaultCoreEnvironment.java
@@ -869,7 +869,7 @@ public class DefaultCoreEnvironment implements CoreEnvironment {
          * a DCP Buffer Acknowledge message is sent to the server to signal producer how much data has been processed.
          * (default value {@value #DCP_CONNECTION_BUFFER_ACK_THRESHOLD}).
          */
-        public Builder dcpConnectionBufferAckThreshold(final int dcpConnectionBufferAckThreshold) {
+        public Builder dcpConnectionBufferAckThreshold(final double dcpConnectionBufferAckThreshold) {
             this.dcpConnectionBufferAckThreshold = dcpConnectionBufferAckThreshold;
             return this;
         }
